refactor(api): extract tools + sessions handlers from system.rs (#3749 3/8)
routes/system.rs continued to span 8 unrelated domains (#3749). Extract
the tools + sessions sub-domain into a sibling module
routes/tools_sessions.rs, following the pattern established by #4452
(pairing) and #4454 (agent_templates).

Move list_tools, get_tool, invoke_tool, list_sessions, get_session,
delete_session, set_session_label, find_session_by_label,
session_cleanup, search_sessions, and the shared PaginationParams
helper. system::router() now mounts /tools/* and /sessions/* (plus the
per-agent /agents/{id}/sessions/by-label/{label} lookup) via
.merge(crate::routes::tools_sessions::router()) -- public route paths
are unchanged.

PaginationParams lives in the new module because the session listing
endpoints are its primary user; system::list_approvals reuses it via
super::tools_sessions::PaginationParams. The const + helper methods
are pub(super) so the cross-module reference compiles without widening
the API surface.

routes/mod.rs gets pub mod tools_sessions + pub use tools_sessions::*
so existing utoipa path bindings in openapi.rs (routes::list_tools,
routes::list_sessions, ...) continue to resolve.

system.rs goes from 5939 -> 5295 lines (-644); new tools_sessions.rs is
674 lines.

Refs #3749 (3 of 8 sub-domain extractions -- remaining: audit_log,
approvals, hooks, commands, batch_resolve).
